penni-chatbot/requirements.txt
langchain-google-vertexai
langgraph
langgraph-checkpoint-postgres
langsmith
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
tenacity
google-cloud-firestore
google-cloud-secret-manager
google-cloud-logging
google-auth
firebase-admin
psycopg2-binary
httpx
orjson
python-json-logger
pytest
pytest-asyncio
